pyproject.toml
[tool.poetry]
name = "ha-greenwichtunnel"
version = "0.1.0"
description = "Home Assistant integration exposing community-sourced lift status for the Greenwich Foot Tunnel"
authors = ["hypercubian <hypercubian@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "custom_components" }]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
# Pin josepy below 2.0 to keep the hass_nabucasa -> acme chain importable.
# acme 2.11 (HA's current pin) still references josepy.ComparableX509 which was
# removed in josepy 2.0. Remove this pin once HA bumps acme.
josepy = "<2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.24"
pytest-cov = "^5.0"
pytest-homeassistant-custom-component = "^0.13.180"
aioresponses = "^0.7"
black = "^24.10"
isort = "^5.13"
flake8 = "^7.1"
flake8-bugbear = "^24.10"
flake8-comprehensions = "^3.15"
flake8-docstrings = "^1.7"
mypy = "^1.13"
pylint = "^3.3"
bandit = "^1.7"
pre-commit = "^4.0"
pyupgrade = "^3.19"
docformatter = "^1.7"
detect-secrets = "^1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["custom_components"]

[tool.mypy]
python_version = "3.13"
strict = true
disallow_untyped_defs = true
disallow_untyped_decorators = true
disallow_any_generics = true
warn_unreachable = true
show_error_codes = true
plugins = []

[[tool.mypy.overrides]]
module = ["homeassistant.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
  "unit: fast unit tests with no I/O",
  "integration: cross-component tests",
  "functional: end-to-end BDD scenarios",
  "slow: long-running tests",
]

[tool.coverage.run]
source = ["custom_components/greenwich_tunnel"]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
]
show_missing = true

[tool.pylint.main]
py-version = "3.13"
jobs = 0

[tool.pylint."messages control"]
disable = ["too-few-public-methods", "duplicate-code"]

[tool.bandit]
exclude_dirs = ["tests"]
